requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
requests==2.32.4
scipy==1.16.0
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tifffile==2025.6.11
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
